requirements.txt
Django==1.7.5
django-braces==1.4.0
django-oauth-toolkit==0.8.0
djangorestframework==2.4.3
--allow-external mysql-connector-python
mysql-connector-python==2.0.3
oauthlib==0.7.2
psycopg2==2.6
six==1.9.0
